+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.2
+
+*  Change default total objects to 40
+*  Change default rotation speed to 0.02
+*  Change default particle size to 30
+*  Change default start from top to false
+*  Now if rotation speed is 0 initial rotation is also 0
+*  Update example project
+
 ## 1.0.1
 
 *  Fix issues in documentation